requirements.txt
streamlit
easyocr
opencv-python-headless
pillow
numpy
speechrecognition
pyaudio
